Get the sprites to layer correctly
We're setting the depth of the sprites to be their y coordinate plus
half their height (`displayHeight` is after scaling). This means they
get layered intuitively based on the position of the bottom of the
sprite.

I've also tweaked the poop creation, triggering the animation, and
delaying the poop creation and sound effect appropriately.